# Conflicts:
#	SEP490/Program.cs
#	fe/app/(defaults)/inventoryslip/service.ts
#	fe/setup/axios.ts

+++ SEP490/Modules/ProductionOrders/DTO/ProductionOutputDto.cs
@@ -6,15 +6,15 @@ namespace SEP490.Modules.ProductionOrders.DTO
         public int ProductId { get; set; }
         public string? ProductName { get; set; }
         public decimal? Amount { get; set; }
-        public int? Done { get; set; }
-        public int? Broken { get; set; }
+        public decimal? Done { get; set; }
+        public decimal? Broken { get; set; }
         public string? ReasonBroken { get; set; }
         public int? ProductionOrderId { get; set; }
     }
 
     public class ReportBrokenOutputDto
     {
-        public int Broken { get; set; }
+        public decimal Broken { get; set; }
         public string ReasonBroken { get; set; } = string.Empty;
     }
