+++ SEP490/Modules/ProductionPlanModule/DTO/ProductionPlanOutputDto.cs
@@ -6,8 +6,8 @@ namespace SEP490.Modules.Production_plans.DTO
         public int ProductId { get; set; }
         public string? ProductName { get; set; }
         public decimal? TotalAmount { get; set; }
-        public int? Done { get; set; }
-        public int? Broken { get; set; }
+        public decimal? Done { get; set; }
+        public decimal? Broken { get; set; }
         public string? BrokenDescription { get; set; }
     }
 }
